+++ NEWS.md
@@ -1,6 +1,8 @@
 ## 2019-09-09 v2.7.0-SNAPSHOT
 * Changed payload of endpoint for updating parsed records from ParsedRecordCollection to RecordCollection
 * Added order number to the record
+* Updated raml-module-builder version.
+* Updated data-import-utils version.
 
 ## 2019-09-09 v2.6.0
 * Fixed security vulnerability with jackson databind