+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added order number to the record
 * Updated raml-module-builder version.
 * Updated data-import-utils version.
+* Deleted outdated index definition for foreign key.
 
 ## 2019-09-09 v2.6.0
 * Fixed security vulnerability with jackson databind