+++ NEWS.md
@@ -1,4 +1,4 @@
-## 2019-09-09 v2.7.0-SNAPSHOT
+## 2019-11-04 v2.7.0
 * Changed payload of endpoint for updating parsed records from ParsedRecordCollection to RecordCollection
 * Added order number to the record
 * Fixed sorting records by order number